abstracting ExcelRenderTable
